Ex4/requirements.txt
numpy
base_module
__future__
abc
typing 
pandas
sklearn.metrics
utils
plotly.graph_objects